% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{available_experiment_dates}
\alias{available_experiment_dates}
\title{serves available experiment start dates of the dataset}
\usage{
available_experiment_dates(data)
}
\arguments{
\item{data}{the varioscan data tibble}
}
\description{
serves available experiment start dates of the dataset
}
