% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{available_extracts}
\alias{available_extracts}
\title{serves available extracts of the dataset for the given experiments}
\usage{
available_extracts(data, experiment_dates)
}
\arguments{
\item{data}{the varioscan data tibble}

\item{the}{experiments}
}
\description{
serves available extracts of the dataset for the given experiments
}
