% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_analysis.R
\name{build_model}
\alias{build_model}
\title{Builds the model using nls}
\usage{
build_model(
  series,
  series_data,
  dilution_seq,
  dependent_var = "yield",
  default_power = 3,
  nls_trace = FALSE
)
}
\description{
Builds the model using nls
}
