% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_analysis.R
\name{calculate_series_meta_data}
\alias{calculate_series_meta_data}
\title{Calculates IC50 and IC90}
\usage{
calculate_series_meta_data(current_key, the_model, default_power, data_key)
}
\description{
Calculates IC50 and IC90
}
