% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_varioscan.R
\name{check_remote_for_new_datasets}
\alias{check_remote_for_new_datasets}
\title{Reads remote data store and compares to the data collection already present}
\usage{
check_remote_for_new_datasets(all_experiment_dates)
}
\description{
Assumes that the directory has a .htaccss file with this content:
DirectoryIndex index
IndexOptions Type=text/plain SuppressSize
}
