% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_analysis.R
\name{create_wide_data}
\alias{create_wide_data}
\title{Pivot wider and mutate to match requirements of the \code{growthcurver} package}
\usage{
create_wide_data(varioscan)
}
\arguments{
\item{varioscan}{the varioscan data in long format}
}
\description{
This function will export the background_corrected data only.
}
