% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_analysis.R
\name{determine_growth_params}
\alias{determine_growth_params}
\title{Uses \code{growthcurver::SummarizeGrowth} to determine growth}
\usage{
determine_growth_params(wide_format_data)
}
\description{
Uses \code{growthcurver::SummarizeGrowth} to determine growth
}
