% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_analysis.R
\name{do_growth_analysis}
\alias{do_growth_analysis}
\title{Perform growth analysis on a varioscan data set}
\usage{
do_growth_analysis(varioscan)
}
\arguments{
\item{varioscan}{the varioscan data in long format}
}
\description{
Perform growth analysis on a varioscan data set
}
