% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{exclude_data}
\alias{exclude_data}
\title{not exported helper function to exclude specific wells or columns}
\usage{
exclude_data(data, exclude)
}
\description{
not exported helper function to exclude specific wells or columns
}
