% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter Varioscan data}
\usage{
filter_data(
  data,
  extracts = "all",
  strains = "all",
  replicates = "all",
  experiment_dates = "all",
  exclude = "none"
)
}
\arguments{
\item{data}{the dataframe holding all data}

\item{extracts}{which extracts to include. Defaults to all extracts.}

\item{strains}{the bacterial strains to include. Defaults to all strains.}

\item{replicates}{the replicates to include. Defaults to all replicates.}

\item{experiment_dates}{the start date values to include}

\item{exclude}{a list indicating which series/dilution/replicate should be excluded. Defaults to "none".
See details.}
}
\value{
filtered data with same columns
}
\description{
General-purpose function to filter Varioscan data in long format on different column values.
Assumes these columns exist:
\code{
  [1] "dilution"       "time"           "series"         "replicate"      "OD"
  [6] "duration"       "start_date"     "strain"         "extract"        "date_extracted"
  [11] "medium"
}
}
\details{
\code{exclude} The \code{exclude} parameter can be used to specifically exclude a
replicate of a series or a (set of) dilution(s) of a replicate (of a series)
of a specific experiment of course.
It is intended to use for excluding failed wells from the plate.
It should be provided as a list in this form:\preformatted{exclude = list(
    start_date = "30-1-2020",
    series = "Exp1",
    replicate = "1", #will remove both 1 and 1C
    dilution = c(0.02, 0.01) #if left empty, all dilutions of te replicate will be excluded
)
}

Optionally, a list-of lists can be provided with repeated definitions of items to exclude:\preformatted{exclude = list(
    list(
        start_date = "30-1-2020",
        series = "Exp1",
        replicate = "1", #will remove both 1 and 1C
        dilution = c(0.02, 0.01) #if left empty, all dilutions of the replicate will be excluded
    ),
    list(
        start_date = "30-1-2020",
        series = "Exp2",
        replicate = "2"
        #empty, so all dilutions
    )
)
}
}
\examples{
#filter replicates
filter_data(all_data, replicates = c("1C", "2C"))

# filter replicates and extract
filter_data(all_data, replicates = c("1C", "2C"), extracts = "Tulip stem")

filter replicates and lower date
filter_data(all_data, replicates = c("1C", "2C"))

}
