% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_checkbox_table.R, R/shiny_helpers.R
\name{get_clickable_plate_layout}
\alias{get_clickable_plate_layout}
\title{Adds glyphicons to a given plate layout}
\usage{
get_clickable_plate_layout(ui_glyphs, plate_layout)

get_clickable_plate_layout(ui_glyphs, plate_layout)
}
\description{
Adds glyphicons to a given plate layout

Adds glyphicons to a given plate layout
}
