% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{get_new_glyph}
\alias{get_new_glyph}
\title{returns new glyph depending on the glyph that was clicked}
\usage{
get_new_glyph(clicked_glyph, ui_glyphs)
}
\description{
returns new glyph depending on the glyph that was clicked
}
