% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{get_selected_wells}
\alias{get_selected_wells}
\title{serves a list of selected wells, to be used for downstram filtering using the filter_data function}
\usage{
get_selected_wells(wells, start_date_of_exp)
}
\description{
serves a list of selected wells, to be used for downstram filtering using the filter_data function
}
