% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_varioscan.R
\name{load_selected_experiment}
\alias{load_selected_experiment}
\title{Loads a single experiment from Rda file (from package package data)}
\usage{
load_selected_experiment(experiment_date_single)
}
\description{
Loads a single experiment from Rda file (from package package data)
}
