% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{message_helper}
\alias{message_helper}
\title{helper for creating messages}
\usage{
message_helper(message_text, vector)
}
\arguments{
\item{message_text}{the prepended massege text}

\item{vector}{the vector of elements to message, with separating symbols}
}
\description{
helper for creating messages
}
