% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varioscan_class.R
\name{metadata}
\alias{metadata}
\title{Constructor for the metadata object that will be part of the varioscan data}
\usage{
metadata(
  date_started,
  dilutions = c(0.02, 0.01, 0.005, 0.0025, 0.0013, 6e-04, 3e-04, 0),
  experiment_name = NA
)
}
\description{
Constructor for the metadata object that will be part of the varioscan data
}
