% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_analysis.R
\name{model_dose_response}
\alias{model_dose_response}
\title{Models the yield as a function of dilution}
\usage{
model_dose_response(
  growth_params,
  dependent_var = "AUC_l",
  data_key = "series",
  nls_trace = FALSE
)
}
\arguments{
\item{growth_params}{the tibble with growth parameters (a result of \code{do_growth_analysis()})}

\item{nls_trace}{when TRUE gives the tarce of the nls function}

\item{dependent}{the dependent variable, either one of \code{yield} and \code{auc_l}}
}
\description{
Models the yield as a function of dilution
}
