% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growthcurves.R
\name{plot_growthcurves}
\alias{plot_growthcurves}
\title{Plots varioscan data in growth curves}
\usage{
plot_growthcurves(varioscan_data, only_average = FALSE)
}
\arguments{
\item{varioscan_data}{all data in long format}

\item{only_average}{whether only the average of the three series should be displayed}
}
\description{
Plots varioscan data in growth curves
}
