% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{prepare_stats_data_download}
\alias{prepare_stats_data_download}
\title{prepares the growth parameters dataframe, as}
\usage{
prepare_stats_data_download(growth_params)
}
\description{
prepares the growth parameters dataframe, as
}
