% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_varioscan.R
\name{read_experiment_data}
\alias{read_experiment_data}
\title{Reads experiment names to be used in the app}
\usage{
read_experiment_data(xlsx_file)
}
\arguments{
\item{xlsx_file}{An excel file with two columns: \code{file_name} and \code{experiment_name}}
}
\value{
a tibble with 3 columns
}
\description{
Reads experiment names to be used in the app
}
