% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_varioscan.R
\name{read_varioscan}
\alias{read_varioscan}
\title{Reads Varioscan data from Excel}
\usage{
read_varioscan(
  xlsx_file,
  dilutions = c(0.02, 0.01, 0.005, 0.0025, 0.0013, 6e-04, 3e-04, 0),
  experiments = c("Exp1", "Exp2", "Exp3"),
  data_series = paste0(rep(experiments, each = 4), "_", c("1", "2", "3", "C")),
  time_interval = 10
)
}
\arguments{
\item{xlsx_file}{A varioscan data file}

\item{dilutions}{the dilutions used in the subsequent rows for dose-response studies}

\item{experiments}{the 3 experiments (will end up as labels in the long format).}

\item{time_interval}{the time interval, in minutes, between each reading}
}
\value{
all data in long format
}
\description{
Reads varioscan data from a given excel file. Assumes data is arranged in 96-well format (8 rows and 12 columns). It also assumes a triplicate experimental setup: 3 curves in triplo, each with a control at the fourth column. If the "General_info" tab does not contain metadata at the correct location, the \code{experiments} argument will be used for the labeling
}
