% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth_statistics.R
\name{remove_extreme_yield_outliers}
\alias{remove_extreme_yield_outliers}
\title{removes extreme outliers with respect to yield
because of plot corruption}
\usage{
remove_extreme_yield_outliers(growth_params_tibble, z_score_cutoff = 3)
}
\description{
removes extreme outliers with respect to yield
because of plot corruption
}
