% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{shiny_app}
\alias{shiny_app}
\title{starts the shiny app}
\usage{
shiny_app(launch_browser = FALSE, port = 3838)
}
\arguments{
\item{launch_browser}{logi flagging whether to launch the default browser of
this device and open the app in it.}

\item{port}{the port to listen to.}
}
\description{
The Shiny app is a GUI wrapper around the main functionality of this package
}
