% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helper_functions.R
\name{available_extracts}
\alias{available_extracts}
\title{serves available extracts of the dataset}
\usage{
available_extracts(data)
}
\arguments{
\item{data}{the varioscan data tibble}
}
\description{
serves available extracts of the dataset
}
