% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{available_strains}
\alias{available_strains}
\title{serves available strains within the dataset}
\usage{
available_strains(data)
}
\arguments{
\item{data}{the varioscan data tibble}
}
\description{
serves available strains within the dataset
}
