% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter Varioscan data}
\usage{
filter_data(
  data,
  extracts = "all",
  strains = "all",
  replicates = "all",
  lower_date = lubridate::today() - 7,
  upper_date = lubridate::today()
)
}
\arguments{
\item{data}{the dataframe holding all data}

\item{extracts}{which extracts to include. Defaults to all extracts.}

\item{strains}{the bacterial strains to include. Defaults to all strains.}

\item{replicates}{the replicates to include. Defaults to all replicates.}

\item{lower_date}{Lubridate Date object: the lower date (start_date of experiment) to include, and later. Defaults to a week ago from the current day.}

\item{upper}{Lubridate Date object: the upper date (inclusive) to include, and before. Defaults to the current day.}
}
\description{
General-purpose function to filter Varioscan data in long format on different column values.
Assumes these columns exist:
\code{
  [1] "dilution"       "time"           "series"         "replicate"      "OD"
  [6] "duration"       "start_date"     "strain"         "extract"        "date_extracted"
  [11] "medium"
}
}
\examples{
#filter replicates
filter_data(all_data, replicates = c("1C", "2C"))
# filter replicates and extract
filter_data(all_data, replicates = c("1C", "2C"), extracts = "Tulip stem")
filter replicates and lower date
filter_data(all_data, replicates = c("1C", "2C"), lower_date = lubridate::dmy("2-3-2020"))
}
