% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{get_plate_layout}
\alias{get_plate_layout}
\title{returns the plate layout belonging to the given experiment}
\usage{
get_plate_layout(selected_experiment)
}
\description{
returns the plate layout belonging to the given experiment
}
