% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growthcurves.R
\name{plot_growthcurves}
\alias{plot_growthcurves}
\title{Plots varioscan data in growth curves}
\usage{
plot_growthcurves(varioscan_data, type = "replicates")
}
\arguments{
\item{varioscan_data}{all data in long format}

\item{type}{a character, one of \code{c("replicates", "avg", "limits")}. See details}
}
\description{
Plots varioscan data in growth curves
}
\details{
\itemize{
\item \code{type} The \code{type} parameter can be one of \code{c("replicates", "avg", "limits")}
\itemize{
\item \code{replicates} the individual replicates will be displayed
\item \code{avg} only the average of the three series will be displayed
\item \code{limits} both average and maximum and minimum will be displayed
}
}
}
